LC-maximum-value-of-an-ordered-triplet-i [Runtime: 3 ms; Memory: 21.5 MB]

+++ array/maximum-value-of-an-ordered-triplet-i.cpp
@@ -6,7 +6,7 @@ public:
         for(int i=0; i<n; i++){
             for(int j=i+1; j<n; j++){
                 for(int k=j+1; k<n; k++){
-                    long long currAns = (nums[i]-nums[j])*nums[k];
+                    long long currAns = ((long long)nums[i]-(long long)nums[j])*(long long)nums[k];
                     ans = max(ans, currAns);
                 }
             }
